Change name of priority object and adjust packet listener
Change the name of ListenerPriority to ProtocolPriority to more
differentiate from other types of listeners.

Adjust the packet listener to remove the option for a listener to
receive a cancelled packet. The protocol manager was never setup to
handle cancelled packets and there is not much reason to continue a
packet down the pipeline once it has been cancelled since it cannot
be uncancelled.

Signed-off-by: Mark <33119mark@gmail.com>